clean up code - sort out the interfaces/schemas, move utility functions into own place, update slight differences between FE/BE
