+++ src/schemas/index.ts
@@ -1,0 +1,7 @@
+export { pingSchema } from './ping';
+export {
+	locationSchema,
+	locationListSchema,
+	locationQueryParamsSchema,
+} from './location';
+export { newSpecialRequestSchema, newSpecialResponseSchema } from './special';
