+++ src/schemas/ping.ts
