+++ src/services/api/methods.ts
@@ -3,15 +3,15 @@ import {
 	pingSchema,
 	locationListSchema,
 	locationQueryParamsSchema,
-	newSpecialResposneSchema,
-} from 'types/schemas';
+	newSpecialResponseSchema,
+} from 'schemas';
 import Endpoints from './endpoints';
 import { zodGet, zodPost } from './zod-methods';
 import { validateAsync } from './validate';
 import type {
 	PingResponse,
-	LocationListResponse,
-	NewSpecial,
+	FormattedLocation,
+	newSpecialRequest,
 	LocationsQueryParams,
 } from 'types';
 
@@ -23,19 +23,19 @@ export const getLocations = (params: LocationsQueryParams) => {
 	const searchParams = new URLSearchParams();
 	Object.entries(params).forEach(([key, value]) => {
 		if (!isNil(value) && value !== '') {
-			searchParams.append(snakeCase(key), value);
+			searchParams.append(snakeCase(key), value as string);
 		}
 	});
 	return validateAsync(locationQueryParamsSchema, params).then(() =>
-		zodGet<LocationListResponse>(
+		zodGet<FormattedLocation[]>(
 			Endpoints.locations(searchParams.toString()),
 			locationListSchema,
 		),
 	);
 };
 
-export const postNewSpecial = (data: NewSpecial) => {
+export const postNewSpecial = (data: newSpecialRequest) => {
 	// eslint-disable-next-line no-console
 	console.log('new data', data);
-	return zodPost(Endpoints.specials, data, newSpecialResposneSchema);
+	return zodPost(Endpoints.specials, data, newSpecialResponseSchema);
 };
