--- src/types/schemas/index.ts
@@ -1,16 +1,0 @@
-export { pingSchema, type PingResponse } from './testingSchema';
-export {
-	locationSchema,
-	locationListSchema,
-	locationQueryParamsSchema,
-	type FormattedLocation,
-	type LocationListResponse,
-} from './locationSchema';
-export {
-	specialSchema,
-	newSpecialSchema,
-	newSpecialResposneSchema,
-	type SpecialResponse,
-	type SpecialListResponse,
-	type NewSpecial,
-} from './specialSchema';
